+++ src/e2e/fixture/metadata.rs
@@ -65,6 +65,8 @@ pub enum FixtureDocsOperation {
         #[serde(default)]
         fields: Vec<String>,
         #[serde(default)]
+        display: bool,
+        #[serde(default)]
         optional: bool,
     },
 }